+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and what APIs have changed, if applicable.
 ## [Unreleased]
 - Add fluent client api for subresources
 - Update fluent client APIs to include projection mask as input parameter.
+- Update projection mask builder APIs to support updating the mask objects.
 
 ## [29.17.0] - 2021-03-23
 - Implement D2 cluster subsetting.